requirements.txt
numpy>=1.21.0
pandas>=1.4.0
requests>=2.28.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
hyperliquid-python-sdk>=0.1.0
eth-account>=0.8.0
tqdm>=4.65.0
